requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
requests
